Discofreezer Functional Game Update
The game works. More or less. There are no side buttons, nor are there victory or loss conditions. There is a weird bug that involves dropped inputs that are still stored, not sure why tbh. It may be out of (our) scope, but not necessarily unfixable. Almost all changes are to the main. The "mut" attribute of the Cell class has been modified slightly. 0 represents immutable, 1 is submitted, 2 is neither. Both 1 and 2 are mutable, but 1 is only mutable if cleared first. Checking for no mut == 2 in board is an easy way to check for board completion.
The selection part of the draw method of the Cell class is fully implemented and working. I don't really remember how I did it, but it's vaguely jerryrigged. Careful when modifying. Anyways, almost done!!!